+++ uhabits-core/src/commonTest/kotlin/org/isoron/uhabits/core/models/StreakListTest.kt
@@ -64,4 +64,17 @@ class StreakListTest : BaseUnitTest() {
         assertEquals(1, best.size)
         assertEquals(1, best[0].length)
     }
+
+    @Test
+    fun testSkipDoesNotBreakNumericalAtLeastStreak() {
+        habit = fixtures.createEmptyNumericalHabit(NumericalHabitType.AT_LEAST)
+        habit.frequency = Frequency.DAILY
+        habit.originalEntries.add(Entry(today, 2000))
+        habit.originalEntries.add(Entry(today.minus(1), Entry.SKIP))
+        habit.originalEntries.add(Entry(today.minus(2), 2000))
+        habit.recompute()
+        val best = habit.streaks.getBest(1)
+        assertEquals(1, best.size)
+        assertEquals(3, best[0].length)
+    }
 }